+++ src/common/enums/anti-gradient-mode.enum.js
@@ -1,7 +1,0 @@
-var AntiGradientMode = Object.freeze({
-  Disabled: 0,
-  Lax: 1,
-  Strict: 2
-});
-
-export default AntiGradientMode;