+++ src/ext/conf/ExtConfPatches.js
@@ -1,8 +1,8 @@
 // How to use:
 // version: {ExtensionConf object, but only properties that get overwritten}
-import Stretch from '../../common/enums/stretch.enum';
-import ExtensionMode from '../../common/enums/extension-mode.enum';
-import VideoAlignment from '../../common/enums/video-alignment.enum';
+import StretchType from '../../common/enums/StretchType.enum';
+import ExtensionMode from '../../common/enums/ExtensionMode.enum';
+import VideoAlignmentType from '../../common/enums/VideoAlignmentType.enum';
 
 const ExtensionConfPatch = [
   {
@@ -267,7 +267,7 @@ const ExtensionConfPatch = [
             label: '4:3 stretch (src)',
             cmd: [{
               action: 'set-stretch',
-              arg: Stretch.FixedSource,
+              arg: StretchType.FixedSource,
               customArg: 1.33,
             }],
             scopes: {
@@ -284,7 +284,7 @@ const ExtensionConfPatch = [
             label: '16:9 stretch (src)',
             cmd: [{
               action: 'set-stretch',
-              arg: Stretch.FixedSource,
+              arg: StretchType.FixedSource,
               customArg: 1.77,
             }],
             scopes: {
@@ -309,8 +309,8 @@ const ExtensionConfPatch = [
         autoar: ExtensionMode.Enabled,
         autoarFallback: ExtensionMode.Enabled,
         override: true,                  // ignore value localStorage in favour of this
-        stretch: Stretch.Default,
-        videoAlignment: VideoAlignment.Default,
+        stretch: StretchType.Default,
+        videoAlignment: VideoAlignmentType.Default,
         keyboardShortcutsEnabled: ExtensionMode.Default,
         DOM: {
           player: {
@@ -339,8 +339,8 @@ const ExtensionConfPatch = [
           autoar: ExtensionMode.Enabled,     
           override: false,
           type: 'community',
-          stretch: Stretch.Default,
-          videoAlignment: VideoAlignment.Default,
+          stretch: StretchType.Default,
+          videoAlignment: VideoAlignmentType.Default,
           keyboardShortcutsEnabled: ExtensionMode.Default,
           arPersistence: true,              // persist aspect ratio between different videos
           autoarPreventConditions: {        // prevents autoar on following conditions