+++ CHANGELOG.md
@@ -8,13 +8,23 @@
 * Settings page looks ugly af right now. Maybe fix it some time later
 * other bug fixes
 
-## v5.0 (planned major)
+## v6.0 (planned major)
 
  * WebGL autodetection
  * in-player GUI
  * Fix UI logger
 
+## v5.x (next major)
+
+* Migrate main scripts to typescript (vue is currently not included)
+
 ## v4.x (current major)
+
+### v4.5.3
+
+* Provides workaround for the fullscreen stretching bug Chrome 88 (or a recent Windows 10 update) introduced for nVidia users using hardware acceleration on Windows 10. In order to mitigate this bug, Ultrawidify needs to keep a 5-10 px wide black border while watching videos in full screen. This bug is also present in Edge.
+* **[4.5.3.1]** Fixed letterbox misalignment binding in settings (#134)
+* **[4.5.3.2]** Fixed false 'autodetection not supported' notifications.
 
 ### v4.5.2